#83 Updated comments for all examples
