+++ examples/Example.LightBDD.MsTest2.UWP/Features/Customer_journey.Steps.cs
@@ -1,9 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Example.Domain.Helpers;
 using LightBDD.Framework;
-using LightBDD.Framework.Scenarios.Basic;
-using LightBDD.Framework.Scenarios.Contextual;
-using LightBDD.Framework.Scenarios.Extended;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 #pragma warning disable 1998
@@ -12,45 +9,6 @@ namespace Example.LightBDD.MsTest2.UWP.Features
 {
     public partial class Customer_journey
     {
-        private async Task<CompositeStep> Then_customer_should_receive_order_email()
-        {
-            return CompositeStep.DefineNew()
-                .WithContext<MailBox>()
-                .AddAsyncSteps(
-                    x => x.Then_customer_should_receive_invoice(),
-                    x => x.Then_customer_should_receive_order_confirmation())
-                .Build();
-        }
-
-        private async Task<CompositeStep> When_customer_pays_for_products_in_basket()
-        {
-            return CompositeStep.DefineNew()
-                .AddSteps(
-                    When_customer_requests_to_pay,
-                    Then_payment_should_be_successful)
-                .Build();
-        }
-
-        private async Task<CompositeStep> When_customer_adds_products_to_basket()
-        {
-            return CompositeStep.DefineNew().AddAsyncSteps(
-                    _ => Given_product_is_in_stock("wooden desk"),
-                    _ => When_customer_adds_product_to_the_basket("wooden desk"),
-                    _ => Then_the_product_addition_should_be_successful())
-                .Build();
-        }
-
-        private async Task<CompositeStep> Given_customer_is_logged_in()
-        {
-            return CompositeStep.DefineNew().AddSteps(
-                    Given_the_user_is_about_to_login,
-                    Given_the_user_entered_valid_login,
-                    Given_the_user_entered_valid_password,
-                    When_the_user_clicks_login_button,
-                    Then_the_login_operation_should_be_successful)
-                .Build();
-        }
-
         #region Emails
         private class MailBox
         {
