+++ examples/Example.LightBDD.MsTest2.UWP/Features/Invoice_feature.cs
@@ -5,6 +5,13 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Example.LightBDD.MsTest2.UWP.Features
 {
+    /// <summary>
+    /// This feature class shows presents an usage of extended scenario syntax that allows to call parameterized steps.
+    /// It also presents usage of LabelAttribute, ScenarioCategoryAttribute and FeatureDescriptionAttribute.
+    /// 
+    /// More information on extended syntax can be found here: https://github.com/LightBDD/LightBDD/wiki/Scenario-Steps-Definition#extended-scenarios
+    /// while attribute usage is described here: https://github.com/LightBDD/LightBDD/wiki/Tests-Structure-and-Conventions
+    /// </summary>
     [TestClass]
     [FeatureDescription(
 @"In order to pay for products
