+++ examples/Example.LightBDD.MsTest2/Features/Basket_feature.cs
@@ -6,6 +6,16 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Example.LightBDD.MsTest2.Features
 {
+    /// <summary>
+    /// This feature class presents that it is possible to use basic scenario syntax with step methods of void or async void signature.
+    /// Scenarios use here the RunScenarioActionsAsync() method that allows mixing void and async void steps.
+    /// 
+    /// As using async void methods is generally not recommended practice (and LightBDD offers ways to handle Task methods),
+    /// LightBDD properly handles async void steps, waiting for all pending tasks to finish, before proceeding to the next step.
+    /// 
+    /// Such way of writing tests may be useful when async steps has to be introduced to existing tests that uses only synchronous step methods.
+    /// More details can be found here: https://github.com/LightBDD/LightBDD/wiki/Scenario-Steps-Definition#asynchronous-scenario-step-execution
+    /// </summary>
     [TestClass]
     [FeatureDescription(
 @"In order to buy products
@@ -27,7 +37,7 @@ I want to add products to basket")]
         }
 
         /// <summary>
-        /// This test presents how LightBDD treats tests with Inconclusive / Ignore asserts
+        /// This test presents how LightBDD treats tests with Inconclusive / Ignore assertions.
         /// </summary>
         [Scenario]
         [Label("Ticket-7")]
