+++ examples/Example.LightBDD.MsTest2/LightBddIntegration.cs
@@ -6,6 +6,12 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Example.LightBDD.MsTest2
 {
+    /// <summary>
+    /// This is a way to enable LightBDD - MsTest integration.
+    /// It is required to do it in all assemblies with LightBDD scenarios.
+    /// The method with [AssemblyInitialize] attribute has to call LightBddScope.Initialize() to initialize LightBDD.
+    /// The method with [AssemblyCleanup] attribute has to call LightBddScope.Cleanup() to finalize LightBDD execution for that assembly and generate reports.
+    /// </summary>
     [TestClass]
     public class LightBddIntegration
     {
@@ -14,6 +20,11 @@ namespace Example.LightBDD.MsTest2
         [AssemblyCleanup]
         public static void Cleanup() { LightBddScope.Cleanup(); }
 
+        /// <summary>
+        /// This method allows to customize LightBDD behavior.
+        /// The code below configures LightBDD to produce also a plain text report after all tests are done.
+        /// More information on what can be customized can be found on wiki: https://github.com/LightBDD/LightBDD/wiki/LightBDD-Configuration#configurable-lightbdd-features 
+        /// </summary>
         private static void OnConfigure(LightBddConfiguration configuration)
         {
             configuration
