+++ examples/Example.LightBDD.XUnit2/Features/Contacts_management.cs
@@ -6,18 +6,19 @@ using LightBDD.XUnit2;
 
 namespace Example.LightBDD.XUnit2.Features
 {
-    /* This feature class presents the usage of contextual scenarios that are used in conjunction with extended step format.
-     * With contextual scenarios it is possible to share state between all steps, which is useful in the situation where
-     * sharing state via feature class fields is not desired.
-     * 
-     * To use contextual scenarios, the Runner.WithContext() method has to be used and the context object can be then accessed by step with lambda parameter.
-     * The scenario below shows the recommended usage of context, where given-when-then steps are declared on the context class (it is however not enforced rule).
-     * 
-     * Another feature presented here is that if lambda parameter name is 1 character only, it is ignored in the reports, and the step type (GIVEN/WHEN/THEN/etc)
-     * is inferred from the step method name.
-     * 
-     * More information on contextual scenarios can be found here: https://github.com/LightBDD/LightBDD/wiki/Scenario-Steps-Definition#contextual-scenarios
-     */
+    /// <summary>
+    /// This feature class presents the usage of contextual scenarios that are used in conjunction with extended step format.
+    /// With contextual scenarios it is possible to share state between all steps, which is useful in the situation where
+    /// sharing state via feature class fields is not desired.
+    /// 
+    /// To use contextual scenarios, the Runner.WithContext() method has to be used and the context object can be then accessed by step with lambda parameter.
+    /// The scenario below shows the recommended usage of context, where given-when-then steps are declared on the context class (it is however not enforced rule).
+    /// 
+    /// Another feature presented here is that if lambda parameter name is 1 character only, it is ignored in the reports, and the step type (GIVEN/WHEN/THEN/etc)
+    /// is inferred from the step method name.
+    /// 
+    /// More information on contextual scenarios can be found here: https://github.com/LightBDD/LightBDD/wiki/Scenario-Steps-Definition#contextual-scenarios
+    /// </summary>
     [FeatureDescription(
 @"In order to maintain my contact book
 As an application user
